- added Male-Asset motion source (e.g. Fleshlight) - reworked asset "hit box" calculation for better adjustability - fixed a bug with radius calculation of AssetReferences when up direction is set to "-UP" and the asset has a certain orientation (e.g. x:40,y:90,z:bugged)
