+++ ToySerialController.csproj
@@ -111,11 +111,13 @@
     <Compile Include="src\Device\TCodeDevice.cs" />
     <Compile Include="src\Device\TCodeDevice.UI.cs" />
     <Compile Include="src\Device\IDevice.cs" />
+    <Compile Include="src\MotionSource\AbstractAssetMotion.cs" />
     <Compile Include="src\MotionSource\References\AssetReference.cs" />
     <Compile Include="src\MotionSource\References\DildoReference.cs" />
     <Compile Include="src\MotionSource\References\IMotionSourceReference.cs" />
     <Compile Include="src\MotionSource\References\MaleReference.cs" />
     <Compile Include="src\MotionSource\CompositeMotionSource.cs" />
+    <Compile Include="src\MotionSource\Targets\AssetTarget.cs" />
     <Compile Include="src\MotionSource\Targets\AbstractPersonTarget.cs" />
     <Compile Include="src\MotionSource\Targets\FemaleTarget.cs" />
     <Compile Include="src\MotionSource\AbstractRefreshableMotionSource.cs" />
