+++ src/MotionSource/Targets/AssetTarget.cs
@@ -1,0 +1,8 @@
+﻿
+namespace ToySerialController.MotionSource
+{
+    public class AssetTarget : AbstractAssetMotion, IMotionSourceTarget
+    {
+        public bool Update(IMotionSourceReference reference) => Update();
+    }
+}
